#233 Refactor bulk edit cell renderer/editor components
