+++ src/app/bulk-edit/bulk-edit-editor/cell-renderers/checkbox-cell-renderer/checkbox-cell-renderer.component.html
@@ -1,0 +1,5 @@
+<input
+  type="checkbox"
+  (click)="checkedHandler($event)"
+  [checked]="params.value"
+/>
