--- src/app/bulk-edit/bulk-edit-editor/renderers/checkbox-renderer/checkbox-renderer.component.html
@@ -1,5 +1,0 @@
-<input 
-type="checkbox" 
-(click)="checkedHandler($event)"
-[checked]="params.value"
-/>
