+++ src/app/bulk-edit/bulk-edit.module.ts
@@ -7,8 +7,8 @@ import { AgGridModule } from 'ag-grid-angular';
 import { BulkEditContainerComponent } from './bulk-edit-container/bulk-edit-container.component';
 import { MaterialModule } from '@mdm/modules/material/material.module';
 import { FormsModule } from '@angular/forms';
-import { CheckboxRendererComponent } from './bulk-edit-editor/renderers/checkbox-renderer/checkbox-renderer.component';
-import { DateCellEditorComponent } from './bulk-edit-editor/renderers/date-cell-editor/date-cell-editor.component';
+import { CheckboxRendererComponent } from './bulk-edit-editor/cell-renderers/checkbox-cell-renderer/checkbox-cell-renderer.component';
+import { DateCellEditorComponent } from './bulk-edit-editor/cell-editors/date-cell-editor/date-cell-editor.component';
 import { CalendarModule } from 'primeng/calendar';
 
 
@@ -26,7 +26,10 @@ import { CalendarModule } from 'primeng/calendar';
     MaterialModule,
     CalendarModule,
     FormsModule,
-    AgGridModule.withComponents([CheckboxRendererComponent, DateCellEditorComponent])
+    AgGridModule.withComponents([
+      CheckboxRendererComponent,
+      DateCellEditorComponent
+    ])
   ],
   exports: [
     BulkEditElementSelectComponent,
